requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.7
first==2.0.1              # via pip-tools
flake8-docstrings==1.1.0
flake8-polyfill==1.0.1    # via flake8-docstrings
flake8==3.4.1
mccabe==0.6.1             # via flake8
pip-tools==1.9.0
pycodestyle==2.3.1        # via flake8
pydocstyle==2.0.0         # via flake8-docstrings
pyflakes==1.5.0           # via flake8
pyyaml==3.12
six==1.11.0               # via pip-tools, pydocstyle
snowballstemmer==1.2.1    # via pydocstyle
